main: view now open most current file
Makes more sense to open the last file created rather than potentially
creating a new one with no content.

The searching for the most current file can be vastly improved by
searching further back than just the first of each month and year.

With the current implementation, the file `2024/08/31.md` won't be
found if search for on 2024/09/01, as the search only goes back to the
first of each month. The same applies for years.